+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/request-detail/components/ConfirmModal.jsx
@@ -78,7 +78,7 @@ export const ConfirmModal = ({ request, requestModalHeader, handleSubmit, trigge
         validationSchema={PayloadSchema}
         onSubmit={onSubmit}
       >
-        {({ isSubmitting, errors, touched }) => (
+        {({ isSubmitting }) => (
           <>
             <Dimmer active={isSubmitting}>
               <Loader inverted size="large" />
