+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/request-detail/components/index.js
@@ -2,3 +2,4 @@ export { RequestDetail } from './RequestDetail';
 export { SideRequestInfo } from './SideRequestInfo';
 export { ActionButtons } from './ActionButtons';
 export { ConfirmModal } from './ConfirmModal';
+export { MainRequestDetails } from './MainRequestDetails';
